Rename asset "spring-sleuth-modern" to "spring-sleuth"
* this would completely replace the old "spring-sleuth" asset with the new app
