+++ assets/spring-sleuth/src/test/java/hello/HelloControllerTest.java
@@ -1,0 +1,31 @@
+package hello;
+
+import static org.hamcrest.Matchers.containsString;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.web.servlet.MockMvc;
+
+@SpringBootTest
+@AutoConfigureMockMvc
+class HelloControllerTest {
+
+    @Autowired
+    private MockMvc mvc;
+
+    @Test
+    void returnsTraceAndSpanDetailsAndParentFromTraceparent() throws Exception {
+        mvc.perform(get("/")
+                .header("traceparent", "00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01"))
+            .andExpect(status().isOk())
+            .andExpect(content().string(containsString("current span: [Trace:")))
+            .andExpect(content().string(containsString("Span:")))
+            .andExpect(content().string(containsString("parents: [00f067aa0ba902b7]")));
+    }
+}
+
